refactor: store & footer
