+++ package.json
@@ -15,6 +15,7 @@
     "bootstrap": "5.2.2",
     "bootstrap-icons": "1.9.1",
     "connected-react-router": "6.9.3",
+    "dompurify": "2.4.1",
     "firebase": "9.14.0",
     "history": "5.3.0",
     "immer": "9.0.16",
@@ -35,6 +36,7 @@
     "@testing-library/jest-dom": "5.16.5",
     "@testing-library/react": "13.4.0",
     "@testing-library/user-event": "14.4.3",
+    "@types/dompurify": "2.4.0",
     "@types/invariant": "2.2.35",
     "@types/jest": "29.2.2",
     "@types/lodash": "4.14.188",
