+++ src/containers/App/index.tsx
@@ -3,6 +3,7 @@ import { Routes, Route, useLocation, useNavigate } from "react-router-dom";
 import get from "lodash/get";
 
 import routes from '../../utils/routes';
+import { VITE_APP } from '../../utils/constants';
 import {
   connect,
   createStructuredSelector,
@@ -11,7 +12,6 @@ import {
   useInjectSaga,
   searchRoute
 } from '../../utils/services';
-import Context from '../../utils/context';
 
 import {
   makeDataSelector as makeDataSelectorError
@@ -39,6 +39,8 @@ import Footer from "../../components/Footer";
 import Logo from "../../components/Logo";
 
 import image from '../../assets/react.svg'
+
+export const USER_NAME = 'Carlos Camacho'
 
 const App = ({
   error,
@@ -83,31 +85,25 @@ const App = ({
   const logo = <Logo width={120} height={120} image={image} />
 
   return (
-    <Context.Provider
-      value={{
-        menu: data.data,
-        message: data.message
-      }}>
-      <Container fluid className="d-flex flex-nowrap p-0">
-        <Header brand='Nova App' userName='Carlos Camacho' menu={data.data} />
+    <Container fluid className="d-flex flex-nowrap p-0">
+      <Header brand={VITE_APP.APP_NAME} userName={USER_NAME} menu={data.data} />
 
-        <Toast open={toast} message={error} onClose={() => setToast(false)} />
+      <Toast open={toast} message={error} onClose={() => setToast(false)} />
 
-        <Container.Content className={`${data.loading ? 'p-0' : ''}`}>
-          <Loading show={data.loading} />
+      <Container.Content className={`${data.loading ? 'p-0' : ''}`}>
+        <Loading show={data.loading} />
 
-          {!data.loading &&
-            <Routes>
-              {routes.map((route: PropsRoute, index: number) => (
-                <Route key={index} path={route.path} element={route.component} />
-              ))}
-            </Routes>
-          }
-        </Container.Content>
+        {!data.loading &&
+          <Routes>
+            {routes.map((route: PropsRoute, index: number) => (
+              <Route key={index} path={route.path} element={route.component} />
+            ))}
+          </Routes>
+        }
+      </Container.Content>
 
-        <Footer logoLeft={logo} logoRight={logo} copyright="Novopayment Inc. All right reserved." />
-      </Container>
-    </Context.Provider>
+      <Footer logoLeft={logo} logoRight={logo} copyright="&copy 2022 Novopayment Inc. All right reserved." />
+    </Container>
   )
 }
