+++ src/main.tsx
@@ -3,8 +3,8 @@ import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 
-import configureStore from './utils/configureStore';
-import history from './utils/history';
+import configureStore from './utils/store/configureStore';
+import history from './utils/store/history';
 import { VITE_APP } from './utils/constants';
 
 import App from './containers/App';
